Wrote all the logic for the main player_decision_state, and wrote the logic for the main methods that I will use later on for the other states

+++ lib/app/game/do_play_black_jack.dart
@@ -28,7 +28,7 @@ class DoPlayBlackJack {
       }
       stdout.write("Insert Cash-in Value: ");
       String? _input3 = stdin.readLineSync();
-      int? _sessionMoney = int.tryParse(_input3 ?? '');
+      double? _sessionMoney = double.tryParse(_input3 ?? '');
       if (_sessionMoney == null || _sessionMoney <= 0){
       //throw InvalidMoneyAmountException
       }
